fix: 恢复2张速度卡无上限档，哨兵值用 Long.MAX_VALUE
- feederPowerEffectiveFeCap 三档：0张=config值 / 1张=×16(钳int.MAX) / ≥2张=Long.MAX_VALUE
- compat 默认 perPassCap 回 Long.MAX_VALUE（无上限哨兵语义统一）
- config 注释/界面文案/日志三态(config|×16|无上限)同步
- 单轮灌入仍受机器 int 缺口限制（min(gap, cap)），哨兵值安全

+++ src/main/java/com/ae2addon/compat/AppFluxPowerCompat.java
@@ -68,13 +68,13 @@ public final class AppFluxPowerCompat {
     /** 给机器充能（多轮，默认每轮无上限）；兼容旧调用。 */
     public static long feedEnergy(BlockEntity target, Direction side,
             appeng.api.networking.IGrid grid, IActionSource source, int passes) {
-        return feedEnergy(target, side, grid, source, passes, Integer.MAX_VALUE);
+        return feedEnergy(target, side, grid, source, passes, Long.MAX_VALUE);
     }
 
     /** 给机器充能（单轮）；保留原签名兼容旧调用。 */
     public static long feedEnergy(BlockEntity target, Direction side,
             appeng.api.networking.IGrid grid, IActionSource source) {
-        return feedEnergy(target, side, grid, source, 1, Integer.MAX_VALUE);
+        return feedEnergy(target, side, grid, source, 1, Long.MAX_VALUE);
     }
 
     /**
